move in type fixes from other PR
