+++ packages/back-end/src/queryRunners/PastExperimentsQueryRunner.ts
@@ -1,7 +1,7 @@
 import { getValidDate } from "shared/dates";
 import {
   PastExperimentParams,
-  PastExperimentResponse,
+  PastExperimentResponseRows,
   PastExperimentResult,
 } from "../types/Integration";
 import {
@@ -154,7 +154,7 @@ export class PastExperimentsQueryRunner extends QueryRunner<
     });
   }
   private processPastExperimentQueryResponse(
-    rows: PastExperimentResponse
+    rows: PastExperimentResponseRows
   ): PastExperimentResult {
     return {
       experiments: rows.map((row) => {
