+++ packages/back-end/src/services/stats.ts
@@ -7,7 +7,7 @@ import {
 import { MetricInterface } from "../../types/metric";
 import { ExperimentMetricAnalysis, StatsEngine } from "../../types/stats";
 import {
-  ExperimentMetricQueryResponse,
+  ExperimentMetricQueryResponseRows,
   ExperimentResults,
 } from "../types/Integration";
 import {
@@ -29,7 +29,7 @@ export const MAX_DIMENSIONS = 20;
 export async function analyzeExperimentMetric(
   variations: ExperimentReportVariation[],
   metric: MetricInterface,
-  rows: ExperimentMetricQueryResponse,
+  rows: ExperimentMetricQueryResponseRows,
   dimension: string | null = null,
   statsEngine: StatsEngine = DEFAULT_STATS_ENGINE,
   sequentialTestingEnabled: boolean = false,
@@ -155,7 +155,7 @@ export async function analyzeExperimentResults({
 }): Promise<ExperimentReportResults> {
   const metricRows: {
     metric: string;
-    rows: ExperimentMetricQueryResponse;
+    rows: ExperimentMetricQueryResponseRows;
   }[] = [];
 
   let unknownVariations: string[] = [];
@@ -169,7 +169,7 @@ export async function analyzeExperimentResults({
     const data = results.result as ExperimentResults;
 
     unknownVariations = data.unknownVariations;
-    const byMetric: { [key: string]: ExperimentMetricQueryResponse } = {};
+    const byMetric: { [key: string]: ExperimentMetricQueryResponseRows } = {};
     data.dimensions.forEach((row) => {
       row.variations.forEach((v) => {
         Object.keys(v.metrics).forEach((metric) => {
@@ -205,7 +205,7 @@ export async function analyzeExperimentResults({
 
       metricRows.push({
         metric: key,
-        rows: query.result as ExperimentMetricQueryResponse,
+        rows: query.result as ExperimentMetricQueryResponseRows,
       });
     });
   }
